# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GovCore Platform API — Identity Engine
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workflow-versions/{id}:
    get:
      operationId: getWorkflowVersion
      tags: [workflow-versions]
      summary: Get a workflow version, including its full states + transitions graph
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workflow version detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowVersionDetail"
        "404":
          description: Not found
  /workflow-versions/{id}/publish:
    post:
      operationId: publishWorkflowVersion
      tags: [workflow-versions]
      summary: Publish a draft version so new instances can be started from it
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowVersion"
        "404":
          description: Not found

  # ─── WORKFLOW INSTANCES ──────────────────────────────────────────────────────
  /workflow-instances:
    get:
      operationId: listWorkflowInstances
      tags: [workflow-instances]
      summary: List workflow instances
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: workflowDefinitionId
          in: query
          schema:
            type: integer
        - name: resourceType
          in: query
          schema:
            type: string
        - name: resourceId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of workflow instances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowInstance"
    post:
      operationId: startWorkflowInstance
      tags: [workflow-instances]
      summary: Start a new workflow instance from the latest published version of a definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowInstanceInput"
      responses:
        "201":
          description: Workflow instance started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowInstanceDetail"
  /workflow-instances/{id}:
    get:
      operationId: getWorkflowInstance
      tags: [workflow-instances]
      summary: Get a workflow instance, including its current state, tasks, and history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workflow instance detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowInstanceDetail"
        "404":
          description: Not found

  # ─── WORKFLOW TASKS ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    WorkflowState:
      type: object
      required: [id, workflowVersionId, name, code, type, isInitial, isFinal, sortOrder]
      properties:
        id:
          type: integer
        workflowVersionId:
          type: integer
        name:
          type: string
        code:
          type: string
        type:
          type: string
          enum: [draft, review, approved, rejected, completed]
        isInitial:
          type: boolean
        isFinal:
          type: boolean
        sortOrder:
          type: integer
...
    WorkflowTransition:
      type: object
      required: [id, workflowVersionId, name, fromStateId, toStateId]
      properties:
        id:
          type: integer
        workflowVersionId:
          type: integer
        name:
          type: string
        fromStateId:
          type: integer
        toStateId:
          type: integer
        requiredPermission:
          type: ["string", "null"]
...
    WorkflowVersion:
      type: object
      required: [id, workflowDefinitionId, version, isPublished, createdAt]
      properties:
        id:
          type: integer
        workflowDefinitionId:
          type: integer
        version:
          type: integer
        isPublished:
          type: boolean
        publishedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    WorkflowVersionDetail:
      allOf:
        - $ref: "#/components/schemas/WorkflowVersion"
        - type: object
          required: [states, transitions]
          properties:
            states:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowState"
            transitions:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowTransition"

    # ── WORKFLOW INSTANCE ──────────────────────────────────────────────────────
    WorkflowInstance:
      type: object
      required: [id, workflowVersionId, tenantId, resourceType, resourceId, currentStateId, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        workflowVersionId:
          type: integer
        tenantId:
          type: integer
        resourceType:
          type: string
        resourceId:
          type: string
        currentStateId:
          type: integer
        status:
          type: string
          enum: [in_progress, completed, cancelled]
        initiatedBy:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkflowInstanceInput:
      type: object
      required: [workflowDefinitionId, resourceId]
      properties:
        workflowDefinitionId:
          type: integer
        resourceId:
          type: string
        firstAssigneeUserId:
          type: integer
        firstAssigneeRoleId:
          type: integer
    WorkflowInstanceDetail:
      allOf:
        - $ref: "#/components/schemas/WorkflowInstance"
        - type: object
          required: [currentState, tasks, history]
          properties:
            currentState:
              $ref: "#/components/schemas/WorkflowState"
            tasks:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowTask"
            history:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowHistoryEntry"

    # ── WORKFLOW TASK ──────────────────────────────────────────────────────────
    WorkflowTask:
      type: object
      required: [id, workflowInstanceId, stateId, status, createdAt]
      properties:
        id:
          type: integer
        workflowInstanceId:
          type: integer
        stateId:
          type: integer
        assigneeUserId:
          type: ["integer", "null"]
        assigneeRoleId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        resolvedBy:
          type: ["integer", "null"]
...
    WorkflowHistoryEntry:
      type: object
      required: [id, workflowInstanceId, toStateId, action, createdAt]
      properties:
        id:
          type: integer
        workflowInstanceId:
          type: integer
        transitionId:
          type: ["integer", "null"]
        fromStateId:
          type: ["integer", "null"]
        toStateId:
          type: integer
        actorUserId:
          type: ["integer", "null"]
        action:
          type: string
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ── RULES (Book 06) ────────────────────────────────────────────────────────
...